add get route method and package

+++ lib/model/remote/google_maps_service/google_maps_models/route_model.dart
@@ -1,5 +1,5 @@
 class RouteModel {
-  double distanceMeters;
+  int distanceMeters;
   String duration;
   Map<String, dynamic> polyline;
